Exam/requirements.txt
torch
torchvision
torchaudio
pandas
openpyxl
pathlib
wandb
matplotlib
pillow
grad-cam
scikit-learn
seaborn
peft
torchmetrics